+++ Ladybird/Qt/AndroidPlatform.cpp
