+++ Ladybird/Qt/InspectorWidget.cpp
