+++ Ladybird/Qt/InspectorWidget.h
