+++ Ladybird/Qt/ModelTranslator.h
