+++ Ladybird/Qt/SettingsDialog.cpp
