+++ Ladybird/Qt/Tab.cpp
@@ -9,8 +9,8 @@
 #include "ConsoleWidget.h"
 #include "InspectorWidget.h"
 #include "Settings.h"
+#include "StringUtils.h"
 #include "TVGIconEngine.h"
-#include "Utilities.h"
 #include <Browser/History.h>
 #include <LibGfx/ImageFormats/BMPWriter.h>
 #include <LibGfx/Painter.h>
