+++ Ladybird/Qt/WebSocketClientManagerQt.cpp
