+++ Ladybird/Qt/WebSocketClientManagerQt.h
