+++ Ladybird/Qt/WebSocketImplQt.h
