+++ Ladybird/Utilities.h
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2022, Andreas Kling <kling@serenityos.org>
+ * Copyright (c) 2023, Andrew Kaster <akaster@serenityos.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -10,13 +11,9 @@
 #include <AK/Error.h>
 #include <AK/String.h>
 #include <AK/Vector.h>
-#include <QString>
 
-AK::DeprecatedString ak_deprecated_string_from_qstring(QString const&);
-ErrorOr<String> ak_string_from_qstring(QString const&);
-QString qstring_from_ak_deprecated_string(AK::DeprecatedString const&);
-QString qstring_from_ak_string(String const&);
 void platform_init();
+ErrorOr<String> application_directory();
 ErrorOr<Vector<String>> get_paths_for_helper_process(StringView process_name);
 
 extern DeprecatedString s_serenity_resource_root;
