+++ Ladybird/WebDriver/CMakeLists.txt
@@ -8,13 +8,13 @@ set(SOURCES
     main.cpp
 )
 
-qt_add_executable(WebDriver ${SOURCES})
+add_executable(WebDriver ${SOURCES})
 
 target_include_directories(WebDriver PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/..)
 target_include_directories(WebDriver PRIVATE ${CMAKE_CURRENT_BINARY_DIR}/..)
 target_include_directories(WebDriver PRIVATE ${SERENITY_SOURCE_DIR}/Userland)
 target_include_directories(WebDriver PRIVATE ${SERENITY_SOURCE_DIR}/Userland/Services)
-target_link_libraries(WebDriver PRIVATE Qt::Core Qt::Network LibCore LibFileSystem LibGfx LibIPC LibJS LibMain LibWeb LibWebSocket)
+target_link_libraries(WebDriver PRIVATE LibCore LibFileSystem LibGfx LibIPC LibJS LibMain LibWeb LibWebSocket)
 add_dependencies(WebDriver headless-browser)
 if (ANDROID)
     link_android_libs(WebDriver)
