+++ Ladybird/WebDriver/main.cpp
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include "../Utilities.h"
 #include <AK/Platform.h>
+#include <Ladybird/Utilities.h>
 #include <LibCore/ArgsParser.h>
 #include <LibCore/Directory.h>
 #include <LibCore/EventLoop.h>
@@ -14,10 +14,7 @@
 #include <LibCore/System.h>
 #include <LibCore/TCPServer.h>
 #include <LibMain/Main.h>
-#include <QCoreApplication>
 #include <WebDriver/Client.h>
-
-extern DeprecatedString s_serenity_resource_root;
 
 static ErrorOr<pid_t> launch_process(StringView application, ReadonlySpan<char const*> arguments)
 {
@@ -57,9 +54,6 @@ static ErrorOr<pid_t> launch_headless_browser(DeprecatedString const& socket_pat
 
 ErrorOr<int> serenity_main(Main::Arguments arguments)
 {
-    // Note: only creating this to get access to its static methods in Utilities
-    QCoreApplication application(arguments.argc, arguments.argv);
-
     auto listen_address = "0.0.0.0"sv;
     int port = 8000;
