--- Ladybird/ladybird.qrc
@@ -1,8 +1,0 @@
-<!DOCTYPE RCC><RCC version="1.0">
-    <qresource>
-        <file>Icons/ladybird.png</file>
-        <file>Icons/back.tvg</file>
-        <file>Icons/forward.tvg</file>
-        <file>Icons/reload.tvg</file>
-    </qresource>
-</RCC>
